feat: Add UserAvatar component
This commit adds a new component called UserAvatar. The UserAvatar component is responsible for rendering a user's avatar image. It takes in an avatarUrl prop, which specifies the URL of the avatar image to be displayed. If no avatarUrl is provided, a default placeholder image is used. The component also accepts optional props for size and className.

The UserAvatar component uses the next/image package to render the image with the specified dimensions. It applies additional styling classes to achieve a rounded, square aspect ratio, and a background color.

This component will be used to display user avatars throughout the application.